chore: move custom exceptions to exceptions.py (#483)
